third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-div-in-anchor.html

<html>
  <head>
    <script src="../../resources/js-test.js"></script>
    <script src="resources/spatial-navigation-utils.js"></script>
    <script type="application/javascript">

    var resultMap = [
      ["Down", "e1"],
      ["DONE", "DONE"]
    ];

    if (window.testRunner) {
      testRunner.dumpAsText();
      testRunner.overridePreference("WebKitTabToLinksPreferenceKey", 1);
      testRunner.overridePreference("WebKitSpatialNavigationEnabled", 1);
      testRunner.waitUntilDone();
    }

    function runTest()
    {
      // starting the test itself: get to a known place.
      document.getElementById("start").focus();

      initTest(resultMap, testCompleted);
    }

    function testCompleted()
    {
      if (window.testRunner)
        testRunner.notifyDone();
    }

    window.onload = runTest;
    </script>
  </head>
<body id="some-content" xmlns="http://www.w3.org/1999/xhtml" style="padding:20px">
<a href="#" id="start"><div>DivInLinkA</div></a>
<a href="#" id="e1"><div>DivInLinkB</div></a>

<div id="console"></div>
This test is testing that we can navigate on links with divs inside.
</body>
</html>
