third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/fast/spatial-navigation/snav-div-in-anchor-expected.txt

DivInLinkA
DivInLinkB
PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
This test is testing that we can navigate on links with divs inside.
